# swagger.yaml
openapi: 3.0.0
info:
  title: Image Processing Service API
  version: 1.0.0
  description: REST API for user authentication, image upload, and image transformation.
servers:
  - url: http://localhost:3000/api/v1
  - url: /api/v1
paths:
  /users/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created
        '400':
          description: Missing fields
        '409':
          description: Username exists
  /users/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: JWT token returned
        '401':
          description: Invalid credentials
  /images:
    post:
      summary: Upload an image
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Image uploaded
        '400':
          description: No file uploaded
    get:
      summary: List images
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: List of images
        '401':
          description: Unauthorized
  /images/{id}:
    get:
      summary: Get image metadata
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Image metadata
        '404':
          description: Not found
  /images/{id}/transform:
    post:
      summary: Transform image
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                operations:
                  type: object
      responses:
        '200':
          description: Transformed image
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT